General adapter setup, config changes, add punk endpoint
