+++ packages/sources/jpegd/schemas/env.json
@@ -1,9 +1,14 @@
 {
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/example-source-adapter.json",
-  "title": "@chainlink/example-source-adapter env var schema",
-  "required": [],
+  "$id": "https://external-adapters.chainlinklabs.com/schemas/jpegd-adapter.json",
+  "title": "Chainlink External Adapter for JPEG'd",
+  "description": "This adapter allows for querying NFT collection values",
+  "required": ["API_KEY"],
   "type": "object",
-  "properties": {},
+  "properties": {
+    "API_KEY": {
+      "type": "string"
+    }
+  },
   "allOf": [
     {
       "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
