+++ packages/sources/jpegd/src/endpoint/index.ts
@@ -1,1 +1,1 @@
-export * as example from './example'
+export * as punks from './punks'
